requirements.txt
streamlit==1.34.0
Pillow==9.2.0